% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhcartr_predictors.R
\name{bagging_predict}
\alias{bagging_predict}
\title{bagging_predict Make predictions for each test dataset row against each tree.}
\usage{
bagging_predict(mytrees, testx, useIdentity, classify, objectid)
}
\arguments{
\item{mytrees}{A list of all trees.}

\item{testx}{The test dataset, the target variables are in the last column.}

\item{useIdentity}{Whether the training data has been transformed with the householder transform.}

\item{classify}{Default is TRUE. Set TRUE for a classification problem and FALSE for a regression problem.}

\item{objectid}{List of objectids (tree nodes) that will be collapsed in the current tree.}
}
\value{
A matrix of all predictions made from all trees and another matrix with a margin for each tree.
}
\description{
This internal function is used to perform some basic checks on the input dataset. The first check to
fail will stop the model from being instantiated. The following checks are performed:
}
