% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_split.R
\name{best_split_}
\alias{best_split_}
\title{best_split_ finds the best feature column to split on.}
\usage{
best_split_(
  X,
  y,
  most_freq_class,
  split_original,
  n_classes,
  max_features,
  depth
)
}
\arguments{
\item{X}{feature variables to search for the best split.}

\item{y}{target variable.}

\item{most_freq_class}{the most frequent class in the target variable.}

\item{split_original}{boolean to indicate whether to split on original data or reflected data.}

\item{n_classes}{number of classes in the y column}

\item{max_features}{the maximum number of features to use when splitting a node}

\item{depth}{the depth of the current tree.}
}
\value{
a list of the following variables (best_idx, best_thr, best_gini)
}
\description{
This internal function is used to find the feature column that will offer the best
split based on using the Gini index or gini hyperplane index.
}
