% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhcartr_calculate_margin.R
\name{calculate_margin_for_tree}
\alias{calculate_margin_for_tree}
\title{calculate_margin_for_tree() - Calculate the margin for a tree.}
\usage{
calculate_margin_for_tree(preds, actuals, count_classified_correct, total)
}
\arguments{
\item{preds}{Predictions on the test set.}

\item{actuals}{The actuals classes ie. the target variables.}

\item{count_classified_correct}{The number of correctly classified predictions.}

\item{total}{The number of rows in the test set.}
}
\value{
returns the margin for the current tree.
}
\description{
This internal function is called from bagging_predict to calculate the margin on the
current tree using the test set.
}
