% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancer.R
\docType{data}
\name{cancer}
\alias{cancer}
\title{Wisconsin Breast Cancer.}
\format{
A data frame with 683 rows and 10 variables:
\describe{
  \item{Sample code number}{id number}
  \item{Clump Thickness}{1-10}
  \item{Uniformity of Cell Size}{1-10}
  \item{Uniformity of Cell Shape}{1-10}
  \item{Marginal Adhesion}{1-10}
  \item{Single Epithelial Cell Size}{1-10}
  \item{Bare Nuclei}{1-10}
  \item{Bland Chromatin}{1-10}
  \item{Normal Nucleoli}{1-10}
  \item{Mitoses}{1-10}
  \item{Class}{2 for benign, 4 for malignant}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/breast+cancer+wisconsin+(original)}
}
\usage{
data(cancer)
}
\description{
This breast cancer databases was obtained from the University of Wisconsin Hospitals,
Madison from Dr. William H. Wolberg.
}
\details{
Attributes 2 through 10 have been used to represent instances. Each instance has one of 2
possible classes: benign or malignant. Original dataset had 699 samples, 16 were removed
as these contained NA's. Class distribution: Benign: 458 (65.5%), Malignant: 241 (34.5%).
}
\keyword{cancer}
\keyword{datasets}
\keyword{wisconsin}
