% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_min_node_impurity.R
\name{compute_min_node_impurity}
\alias{compute_min_node_impurity}
\title{compute_min_node_impurity.}
\usage{
compute_min_node_impurity(X, y, response, seed, useIdentity = FALSE)
}
\arguments{
\item{X}{The feature variables we will use to train the HHDecisionTree model on.}

\item{y}{The target variable.}

\item{response}{The response parameter is used to specify what type of model to build, either 'classify'
for a classification tree model or 'regressor' for a regression tree model. The default is 'classify'.}

\item{seed}{The random seed to be used to prime the random number generator.}

\item{useIdentity}{The useIdentity parameter when set TRUE will result in hhcartr using the
original training data to find the optimal splits rather than using the reflected data. The
default value is FALSE.}
}
\value{
Returns the "best" value of min_node_impurity.
}
\description{
This function instantiates a HHDecisionTree model, it is used to induce classification or regression
trees depending upon the value of the response parameter. It supports the parameters listed below.
}
