% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhcartr_export_displayTree.R
\name{dispnode}
\alias{dispnode}
\title{dispnode generates DOT statements for all trees in current model.}
\usage{
dispnode(nn, level, pid)
}
\arguments{
\item{nn}{current node of the current tree}

\item{level}{current depth in the current tree}

\item{pid}{the parent-id of the current node in the current tree}
}
\value{
nothing
}
\description{
This internal function generates DOT statements for all trees in the current model. It
is called by function navigate_hash.
}
