% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhcartr_getCoefficients.R
\name{getCoefficients}
\alias{getCoefficients}
\title{getCoefficients Get the coefficents from the house holder matrix for the selected node.}
\usage{
getCoefficients(n_node, fold)
}
\arguments{
\item{n_node}{The internal node number the user wants coefficients for. If no node
number is supplied then coefficients for all internal nodes of the current tree are
returned.}

\item{fold}{The fold of the corresponding tree we want coefficients for.}
}
\value{
A dataframe containing the requested information. NA is returned if the requested
node was either not found or was a leaf node.
}
\description{
This function getCoefficients() returns the coefficients from the appropriate column
of the house holder matrix for the selected internal node or for all internal nodes if
no parameter was specified.
}
