% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\docType{data}
\name{glass}
\alias{glass}
\title{Glass Identification Dataset.}
\format{
A data frame with 214 rows and 9 variables:
\describe{
  \item{Id number}{id number}
  \item{RI}{refractive index}
  \item{Na}{Sodium (unit measurement: weight percent in corresponding oxide, as are attributes 4-10) }
  \item{Mg}{Magnesium}
  \item{Al}{Aluminium}
  \item{Si}{Silicon}
  \item{K}{Potassium}
  \item{Ca}{Calcium}
  \item{Ba}{Barium}
  \item{Fe}{Iron}
  \item{Type of Glass}{(class attribute)  -- 1 building_windows_float_processed
  --2 building_windows_non_float_processed -- 3 vehicle_windows_float_processed
  --4 vehicle_windows_non_float_processed (none in this database) -- 5 containers
  -- 6 tableware -- 7 headlamps}
}
}
\source{
\url{https://archive.ics.uci.edu/ml/datasets/Glass+Identification}
}
\usage{
data(glass)
}
\description{
The study of classification of types of glass was motivated by criminological
investigation. At the scene of the crime, the glass left can be used as evidence...
if it is correctly identified!
}
\details{
From USA Forensic Science Service; 6 types of glass; defined in terms of their oxide
 content (i.e. Na, Fe, K, etc).
}
\keyword{datasets}
\keyword{glass}
