% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhcartr_regressor_find_better_split.R
\name{hhcartr_regressor_find_better_split}
\alias{hhcartr_regressor_find_better_split}
\title{hhcartr_regressor_find_better_split finds the best feature to split on.}
\usage{
hhcartr_regressor_find_better_split(X, y, max_features)
}
\arguments{
\item{X}{feature variables to search for the best split.}

\item{y}{target variable.}

\item{max_features}{the maximum number of features to use when splitting a node}
}
\value{
a list of the following variables (var_idx_, split_, score_)
}
\description{
This internal function is used by the regressor model to find the feature column that will offer
the best split based on sum of squares.
}
