% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhcartr_predictors.R
\name{make_predictions}
\alias{make_predictions}
\title{make_predictions Make predictions on the test dataset.}
\usage{
make_predictions(list_trees, test, useIdentity, classify, objectid)
}
\arguments{
\item{list_trees}{List of all induced trees.}

\item{test}{The test dataset.}

\item{useIdentity}{Whether the training data has been transformed with the householder transform.}

\item{classify}{Default is TRUE. Set TRUE for a classification problem and FALSE for a regression problem.}

\item{objectid}{A list of node numbers that will be 'pruned' ie. when making
predictions if the tree node matches a node in objectid the tree node will be used
to make the prediction rather than traversing any underlying nodes.}
}
\value{
Accuracy, margin and predictions.
}
\description{
This internal function is used to make predictions on the test dataset against all induced trees.
}
