% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{NNode}
\alias{NNode}
\title{NNode returns a tree node.}
\usage{
NNode(
  gini,
  num_samples,
  num_samples_per_class,
  predicted_class,
  parent_node,
  objectid,
  oob_row_indices = NA
)
}
\arguments{
\item{gini}{The gini index of the node.}

\item{num_samples}{The number of samples at this node.}

\item{num_samples_per_class}{A table showing class distribution at this node.}

\item{predicted_class}{If a leaf node, the predicted class of this node.}

\item{parent_node}{The parent node object for this child node.}

\item{objectid}{A unique id for this node object.}

\item{oob_row_indices}{For a root node the out-of-bag indices from original training
set used to create this tree.}
}
\value{
a tree node of type hash.
}
\description{
This internal function is used to create a tree node. Remaining parameters will be initialised
as the tree induction process progresses.
}
