% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhcartr_perform_ccp.R
\name{perform_ccp_driver}
\alias{perform_ccp_driver}
\title{perform_ccp_driver.}
\usage{
perform_ccp_driver(treeobjs)
}
\arguments{
\item{treeobjs}{The current HHCARTR tree that has just been induced contained
within a list() object. CCP pruning will be performed on this tree.}
}
\value{
A dataframe containing information for each subtree.
}
\description{
This function performs CCP pruning on the supplied tree.
}
