% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhcartr_reflect_feature_space.R
\name{reflect_feature_space}
\alias{reflect_feature_space}
\title{reflect_feature_space}
\usage{
reflect_feature_space(
  X_matrix,
  X,
  y,
  most_freq_class,
  n_classes,
  max_features,
  n_features,
  depth,
  colx
)
}
\arguments{
\item{X_matrix}{The current node to be tested for a split.}

\item{X}{The feature variables of the current node being processed.}

\item{y}{The target variable for the corresponding feature variables.}

\item{most_freq_class}{The most frequently occurring class at the current node.}

\item{n_classes}{The number of classes in the target variable.}

\item{max_features}{The maximum number of features to consider in the current split.}

\item{n_features}{The number of feature variables.}

\item{depth}{The depth of the current tree.}

\item{colx}{The eigenvector column to use.}
}
\value{
Returns idxA, thrA, gidxA, X_houseA, newH_A if split is on reflected data, otherwise
        a list containing idx, thr, X, NULL, FALSE is returned when the split is made using the
        original data when the first eigen vector equals the first column of the identity matrix.
}
\description{
This function is in an internal only function. It applies the householder transformation
to the feature data.
}
\details{
The following parameters are supported:
}
