% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhcartr_predictors.R
\name{row_predict}
\alias{row_predict}
\title{row_predict Make predictions for a test dataset row against a tree.}
\usage{
row_predict(xnode, test_row, useIdentity, objectid)
}
\arguments{
\item{xnode}{Root node of the current tree we are making predictions against.}

\item{test_row}{Current row from the test dataset.}

\item{useIdentity}{Whether the training data has been transformed with the householder transform.}

\item{objectid}{A list of node numbers that will be 'pruned' ie. when making
predictions if the tree node matches a node in objectid the tree node will be used
to make the prediction rather than traversing any underlying nodes.}
}
\value{
A prediction for the test dataset row.
}
\description{
This internal function is used to run a test dataset row through a tree to make a prediction.
}
