% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhcartr_reflect_feature_space.R
\name{split_using_original_data}
\alias{split_using_original_data}
\title{split_using_original_data}
\usage{
split_using_original_data(
  X,
  y,
  most_freq_class,
  split_original,
  n_classes,
  max_features,
  depth
)
}
\arguments{
\item{X}{The feature variables of the current node being processed.}

\item{y}{The target variable for the corresponding feature variables.}

\item{most_freq_class}{The most frequently occurring class at the current node.}

\item{split_original}{A flag, TRUE if splitting on original data otherwise FALSE.}

\item{n_classes}{The number of classes in the target variable.}

\item{max_features}{The maximum number of features to consider in the current split.}

\item{depth}{The depth of the current tree.}
}
\value{
returns a list containing the following: idx, thr, X, NULL, FALSE.
}
\description{
This function is in an internal only function. It performs an axis-parallel split on the original
data.
}
\details{
The following parameters are supported:
}
